backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-CORS==4.0.0

# Database and Vector Store
qdrant-client==1.7.0

# Data Processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# File Processing
Pillow==10.0.1
opencv-python==4.8.1.78
python-ffmpeg==2.0.12
moviepy==1.0.3

# Audio Processing
librosa==0.10.1
pydub==0.25.1

# API and Data Validation  
pydantic==2.5.0
marshmallow==3.20.1
werkzeug==3.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
python-multipart==0.0.6

# Development
python-decouple==3.8

